assigning workouts to users

+++ frontend/src/components/WorkoutDetails.js
@@ -10,10 +10,13 @@ const WorkoutDetails = ({ workout }) => {
   const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if(!user){
+      return;
+    }
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
       headers: {
-        " Authorization": `Bearer ${user.token}`,
+         'Authorization': `Bearer ${user.token}`,
       },
     });
     const json = await response.json();
@@ -24,7 +27,7 @@ const WorkoutDetails = ({ workout }) => {
   };
 
   return (
-    <>
+    <>    
       <div className="workout-details">
         <h4>{workout.title}</h4>
         <p>
